requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
boto3==1.17.81
    # via -r requirements.in
botocore==1.20.81
    # via
    #   boto3
    #   s3transfer
funcy==1.16
    # via pyldavis
future==0.18.2
    # via pyldavis
gensim==4.0.1
    # via pyldavis
jinja2==3.0.1
    # via pyldavis
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
joblib==1.0.1
    # via
    #   pyldavis
    #   scikit-learn
markupsafe==2.0.1
    # via jinja2
numexpr==2.7.3
    # via pyldavis
numpy==1.20.3
    # via
    #   gensim
    #   numexpr
    #   pandas
    #   pyarrow
    #   pyldavis
    #   scikit-learn
    #   scipy
pandas==1.2.4
    # via
    #   -r requirements.in
    #   pyldavis
pyarrow==2.0.0
    # via -r requirements.in
pyldavis==3.3.1
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   botocore
    #   pandas
python-dotenv==0.17.1
    # via -r requirements.in
pytz==2021.1
    # via pandas
s3transfer==0.4.2
    # via boto3
scikit-learn==0.24.2
    # via
    #   pyldavis
    #   sklearn
scipy==1.6.3
    # via
    #   gensim
    #   pyldavis
    #   scikit-learn
six==1.16.0
    # via python-dateutil
sklearn==0.0
    # via pyldavis
smart-open==5.1.0
    # via gensim
threadpoolctl==2.1.0
    # via scikit-learn
urllib3==1.26.5
    # via botocore

# The following packages are considered to be unsafe in a requirements file:
# setuptools
